requirements.txt
# Django-based Queue Management Bot
django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.9
python-dotenv>=1.1.1
apscheduler>=3.11.0
aiogram>=3.22.0

# Optional async support
asgiref>=3.8.1